Modify the memory pool and remove the deprecated code.
