+++ cocos/core/geometry/frustum.ts
@@ -29,7 +29,6 @@
  */
 
 import { Mat4, Vec3 } from '../math';
-import { FrustumHandle, FrustumPool, FrustumView, NULL_HANDLE } from '../renderer/core/memory-pools';
 import enums from './enums';
 import { Plane } from './plane';
 
@@ -238,32 +237,5 @@ export class Frustum {
         Plane.fromPoints(this.planes[3], this.vertices[0], this.vertices[4], this.vertices[5]);
         Plane.fromPoints(this.planes[4], this.vertices[2], this.vertices[3], this.vertices[0]);
         Plane.fromPoints(this.planes[0], this.vertices[7], this.vertices[6], this.vertices[5]);
-    }
-}
-
-/**
- * @en
- * Record frustum to shared memory.
- * @zh
- * 记录 frustum 数据到共享内存。并不是每个 frustum 都是需要记录到共享内存的。
- * @param handle The frustum handle
- * @param frstm The frustum object
- */
-export function recordFrustumToSharedMemory (handle: FrustumHandle, frstm: Frustum) {
-    if (!frstm || handle === NULL_HANDLE) {
-        return;
-    }
-
-    const vertices = frstm.vertices;
-    let vertexOffset = FrustumView.VERTICES as const;
-    for (let i = 0; i < 8; ++i) {
-        FrustumPool.setVec3(handle, vertexOffset, vertices[i]);
-        vertexOffset += 3;
-    }
-
-    const planes = frstm.planes;
-    let planeOffset = FrustumView.PLANES as const;
-    for (let i = 0; i < 6; i++, planeOffset += 4) {
-        FrustumPool.setVec4(handle, planeOffset, planes[i]);
     }
 }
